backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.7
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
python-binance==1.0.19
celery==5.4.0
redis==5.0.4